Removed project/pio-build.sbt which was creating issues while building. Fixed few issue come while train

+++ src/main/scala/io/prediction/opennlp/engine/Algorithm.scala
@@ -14,7 +14,7 @@ class Algorithm(val ap: AlgorithmParams)
 
   def predict(model: Model, query: Query): PredictedResult = {
     val sentiment = Sentiment(
-      model.gis.getBestOutcome(model.gis.eval(query.sentence.split(" "))).toInt)
+      model.gis.getBestOutcome(model.gis.eval(query.sentence.split(" "))).toFloat.toInt)
     PredictedResult(sentiment.toString)
   }
